Allow actions to specify if they support private bugs.
* Adds an `allow_private` settings to actions, defaults to False.
* Adds parsing of bugzilla groups for actions to use.

By default this doesn't change existing actions but it allows actions to
opt in to syncing private bugs. A custom action my be written to only
sync depending on the groups a bug is a member of. This will only
support bugs in groups that the JBI bugzilla account is a member of.