Always use the bug retrieved through the bugzilla API for field changes
Webhook requests for private bug creation and modification do not
include the normal set of bug fields (for some reason requests for
new comments do). So to make sure you have the full bug info you must
retrieve the bug through the API and then use that to generate the
correct field changes to make in Jira.

This does drop one behaviour, it looks like JBI attempts to modify the
reporter when the reporter changes for a bug. But to my knowledge it
isn't possible to change the reporter for a bug.