Remove top-level `actions` config key (#84)
I noticed that we had a top-level actions key in project config. This seemed
somewhat unnecessary since there aren't any other config objects besides
actions. That meant that any time we wanted to access actions, we'd need to do
`actions_instance.actions.<whatever>`.

This commit removes the top-level actions key in config and modifies the
Actions Pydantic model to give it a `__root__` property to contain the actions.

It also:
* Uses the `parse_file` method of pydantic-yaml to read config
* add descriptions and contact fields to Action model
This includes the `config/` directory and example config in tests