Align infra/test.sh & infra/lint.sh (#88)
The idea with this commit is that `infra/lint.sh` and `infra/test.sh` are the "source
of truth" for how we want to run linting and tests respectively. We then use
these scripts in `precommit`, `make`, Github Actions, and wherever else we want to
run tests and linting.

For Github Actions, we also now run linting and tests directly in the action
runner, rather than running through Docker. This resulted in a speed-up of the
checks.

Fixes #79, #67

Other changes include:
* Do not install postgresql client
* Install git and use precommit for linting
This should fail if it can't run
* Use development image for tests and lint
* Fix lint as reported by `pre-commit run --all-files`, `black`, and `isort`
* Prepend test commands with `poetry run`
* Ignore missing imports for ruamel
* Update .secrets.baseline to exclude poetry.lock
* Add make commands for format, lint and test
* Remove git check for detect-secrets lint check

Co-authored-by: Mathieu Leplatre <mathieu@mozilla.com>