Refactor some test fixtures / setup (#205)
* Decouple test objects from mocked service side effects

In some of our fixtures, we'd return an example test object, but also specify
the returns of mocked side effects when that fixture was used. This made it
slightly more challenging to modify test fixtures. It also wasn't obvious that
the service mocking was happening when reading tests.

Now, we set the return values of mocked services in tests. This makes tests a
bit more verbose, but also makes them more explicit and obvious.

* Add a tests/fixtures/factories.py module to facilitate easy test object creation

These factory functions aren't perfect, and eventually it would be good to
use something like 1pydantic-factories1, but for now this is a good-enough
approach to easily make different examaple objects in tests.

* Move some action tests from test_configuration.py to test_models.py

Some of the Action tests in test_configuration.py were actually testing the
Action validators, so it seemed more appropriate to have them live in
test_models.py.