Start migrating source code to strict typing
In this commit, we start the process of migrating the project to being
strictly typed with mypy.

To do this, we used mypy config to enable all strict type rules that
apply to the module level (some strict rules can't be applied at the
module level). We also specify a list of modules to apply those rules
against. The goal will be to gradually add more modules to the list
until we've added all of the modules, at which point we will check the
entire source directory with `strict` enabled.