Flip default config value `enabled` to `true` (#359)
One of the config values for each project is `enabled`. This value
controls whether actions are run when a matching webhook is encountered.
At the time of this commit, each project set `enabled: true`. So, in
this commit, we flip the default so that actions are enabled by default
and can optionally be disabled by setting `enabled: false`.

* Update .secrets.baseline