Reduce coupling between services and steps modules (#621)
This commit was partially motivated by PR #619. In it, importing the `steps` 
module into the `models` module to introspect it caused a circular import error.
Instead of opting for a late import, I explored alternative approaches to improve
the situation. This commit introduces the following changes:

- Refactor and clarify the purposes of classes in the services modules.

    Each module now contains two classes: a Client responsible only for handling
 HTTP requests, and a Service that performs high-level operations using the 
Client. These changes enhance class responsibilities clarity and facilitate 
testing by passing fake or mock clients directly to the function for assertions 
instead of patching.

 - Reduces coupling between the services and steps modules.

    Step functions now accept services as arguments instead of implicitly
depending on the service modules. This eliminates the need to directly
import services into the steps module (except for type checking) and
simplifies testing by allowing the use of mock or fake services as arguments
instead of patching.

It's worth noting that if we have to import a model into the steps module in 
the future, we may encounter a similar issue. Some more work to decouple
modules may be useful here.

Other changes include:

* Rename FastAPI dependencies

* Add comments about annotations and TYPE_CHECKING