Add `rid` to logs with `asgi-correlation-id` (#782)
* Add `asgi_correlation_id` package

* Replace our RID middleware with `asgi_correlation_id`'s middleware

* Provide logging outside uvicorn

We now call `logging.dictConfig` in app.py, rather that passing the config 
dict to the `log_config` param of uvicorn. This is so that the configuration
is picked up both in tests and when running the app.

* Avoid using TestClient as context manager

I don't know _why_ this was happening, but trying to use a breakpoint in
test_request_summary_is_logged caused the test to hang. I found that using
the fixture directly (i.e. not as a context manager) made the test work as
expected