Change methods for fetching visible projects and issue types
In v6.3.6, we bumped atlassian-python-api from 3.41.4. That bump
included a [PR](https://github.com/atlassian-api/atlassian-python-api/pull/1270)
which changed the way library fetched projects. This somehow affected
our project when we fetched visible projects and issue types in the
heartbeat. We never fully determined the root cause, but the while loop
in the linked PR was suspect.

In this commit, we make further improvements to the code that's involved
in the heartbeat to make it more efficent.

Note that one of the tradeoffs we made for now was hardcoding the
maximum number of projects that may be configured at 50. This is due to
the way we fetch issue types by project. This is something that we can
fix in the future, but is not a problem that we need to solve for now
since we only have 26 project configured.