Add Dead Letter Queue class, integrate into heartbeat (#918)
This commit adds the classes that allow us to manage our "Dead Letter Queue". See https://github.com/mozilla/jira-bugzilla-integration/pull/775 for motivation.

We also add a basic heartbeat check in this commit to assert that we are able to read from / write to the queue. 

In future commits, we'll introduce a method to process items in the queue and fully integrate the queue into the `/bugzilla_webhook` route.

---------

Co-authored-by: Mathieu Leplatre <mathieu@mozilla.com>
Co-authored-by: Alex Cottner <acottner@mozilla.com>